Reverted docker version number and removed wiseconnect-wifi abstraction layer
