+++ .github/workflows/examples-efr32.yaml
@@ -41,7 +41,7 @@ jobs:
     if: github.actor != 'restyled-io[bot]'
 
     container:
-      image: ghcr.io/project-chip/chip-build-efr32:171
+      image: ghcr.io/project-chip/chip-build-efr32:169
       volumes:
         - "/tmp/bloat_reports:/tmp/bloat_reports"
     steps:
