+++ .github/workflows/release_artifacts.yaml
@@ -64,7 +64,7 @@ jobs:
         runs-on: ubuntu-latest
 
         container:
-            image: ghcr.io/project-chip/chip-build-efr32:171
+            image: ghcr.io/project-chip/chip-build-efr32:169
         steps:
             - name: Checkout
               uses: actions/checkout@v5
