+++ src/platform/silabs/rs911x/rsi_ble_config.h
@@ -21,10 +21,8 @@ extern "C" {
 #endif
 
 #include "rsi_ble_apis.h"
-#if (SLI_SI91X_MCU_INTERFACE | EXP_BOARD)
 #include "rsi_bt_common_apis.h"
 #include "rsi_user.h"
-#endif
 
 #ifdef __cplusplus
 }
@@ -94,18 +92,19 @@ extern "C" {
 #define RSI_BLE_CUSTOM_SERVICE_UUID (0xFFF6)
 #define RSI_BLE_CUSTOM_LEVEL_UUID (0x1FF1)
 
-#if (defined(RSI_M4_INTERFACE) || defined(SLI_SI91X_MCU_INTERFACE))
-#define RSI_BLE_MAX_NBR_ATT_REC (20)
-#define RSI_BLE_MAX_NBR_PERIPHERALS (1)
-#define RSI_BLE_MAX_NBR_SLAVES (3)
-#define RSI_BLE_NUM_CONN_EVENTS (2)
+#ifdef SLI_SI91X_MCU_INTERFACE
+#define RSI_BLE_MAX_NBR_ATT_REC 20
+/* Number of BLE notifications */
+#define RSI_BLE_NUM_CONN_EVENTS     2
+#define RSI_BLE_MAX_NBR_PERIPHERALS 1
 #else
-#define RSI_BLE_MAX_NBR_ATT_REC (80)
-#define RSI_BLE_MAX_NBR_PERIPHERALS (3)
-#define RSI_BLE_MAX_NBR_SLAVES (3)
-#define RSI_BLE_NUM_CONN_EVENTS (20)
-#endif // (defined(RSI_M4_INTERFACE) || defined(SLI_SI91X_MCU_INTERFACE))
+#define RSI_BLE_MAX_NBR_ATT_REC     80
+/* Number of BLE notifications */
+#define RSI_BLE_NUM_CONN_EVENTS     20
+#define RSI_BLE_MAX_NBR_PERIPHERALS 3
+#endif
 
+#define RSI_BLE_MAX_NBR_CENTRALS    1
 #define RSI_BLE_MAX_NBR_ATT_SERV (10)
 
 #define RSI_BLE_GATT_ASYNC_ENABLE (1)
