+++ src/platform/silabs/wifi/BUILD.gn
@@ -133,8 +133,6 @@ source_set("wifi-platform") {
     sources += [
       "${silabs_platform_dir}/wifi/SiWx/WifiInterfaceImpl.cpp",
       "${silabs_platform_dir}/wifi/SiWx/WifiInterfaceImpl.h",
-      "${silabs_platform_dir}/wifi/wiseconnect-interface/WiseconnectWifiInterface.cpp",
-      "${silabs_platform_dir}/wifi/wiseconnect-interface/WiseconnectWifiInterface.h",
       "${wifi_sdk_root}/components/service/network_manager/src/sl_net_for_lwip.c",
       "${wifi_sdk_root}/components/service/network_manager/src/sli_net_common_utility.c",
     ]
