+++ src/platform/silabs/wifi/SiWx/ncp/rs9117.gni
@@ -21,8 +21,6 @@ rs911x_src_plat = [
   "${chip_root}/src/platform/silabs/wifi/SiWx/WifiInterfaceImpl.h",
   "${chip_root}/src/platform/silabs/wifi/SiWx/ncp/sl_si91x_ncp_utility.c",
   "${chip_root}/src/platform/silabs/wifi/SiWx/ncp/sl_board_configuration.h",
-  "${chip_root}/src/platform/silabs/wifi/wiseconnect-interface/WiseconnectWifiInterface.cpp",
-  "${chip_root}/src/platform/silabs/wifi/wiseconnect-interface/WiseconnectWifiInterface.h",
   "${chip_root}/src/platform/silabs/wifi/ncp/spi_multiplex.h",
 
   # TODO : We should be using the file from the Wiseconnect SDK and not our copy of it.
